replace mapWithConcurrency helper with p-limit
findInvalidPlugins bounds concurrency with p-limit. Removes the shared
mapWithConcurrency util and its vortex-api export.

LinkingDeployment is restored to its master state, keeping its own local
concurrency helper: the renderer targets node16/CommonJS, which can't
statically import the ESM-only p-limit, and converting it is out of scope
for this PR anyway.
